// src/app/shared/service-proxies/model/adminUserDetailDto.ts
/**
 * Restaurant API v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminUserDetailDto { 
    id?: number;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    isActive?: boolean;
    emailConfirmed?: boolean;
    roles?: Array<string> | null;
    createdAt?: string;
    customerId?: number | null;
    customerIsActive?: boolean | null;
    customerDeletedAt?: string | null;
    staffId?: number | null;
    staffIsActive?: boolean | null;
    staffDeletedAt?: string | null;
}