// src/app/shared/service-proxies/model/adminUserDto.ts
/**
 * Restaurant API v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminUserDto { 
    id?: number;
    email?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    isActive?: boolean;
    emailConfirmed?: boolean;
    roles?: Array<string> | null;
    createdAt?: string;
}