// src/app/shared/service-proxies/model/adminUserDtoPaginatedList.ts
/**
 * Restaurant API v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminUserDto } from './adminUserDto';


export interface AdminUserDtoPaginatedList { 
    items?: Array<AdminUserDto> | null;
    pageNumber?: number;
    readonly totalPages?: number;
    readonly totalCount?: number;
    pageSize?: number;
    readonly hasPreviousPage?: boolean;
    readonly hasNextPage?: boolean;
}