// src/app/shared/service-proxies/model/createAdminUserRequest.ts
/**
 * Restaurant API v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateAdminUserRequest { 
    email?: string | null;
    password?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    role?: string | null;
}